server: removed v1 reporter code base
